@@ -9,7 +9,7 @@ authors = [{ name = "4MBL" }]
 description = "Script that allows you to extract and scale Minecraft's item and block textures."
 readme = "README.md"
 license = { file = "LICENSE" }
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
