@@ -1,12 +1,14 @@
 from abc import ABC, abstractmethod
+from enum import Enum
 import os
 import re
-from shutil import copytree, rmtree
+from shutil import copytree, rmtree, copyfile
 from PIL import Image as pil_image  # type: ignore
 from forfiles import image, file as f  # type: ignore
+
 from .. import texts
 from ..file import mk_dir
-from ..options import DEFAULTS, EditionType, VersionType
+from ..options import DEFAULTS, EditionType, TextureOptions, VersionType
 from ..texts import tabbed_print
 
 REGEX_BEDROCK_RELEASE = r'^v1\.[0-9]{2}\.[0-9]{1,2}\.[0-9]{1,2}$'
@@ -18,21 +20,42 @@ REGEX_JAVA_RC = r'^[0-9]\.[0-9]+\.?[0-9]+-rc[0-9]?$'
 REGEX_JAVA_RELEASE = r'^[0-9]\.[0-9]+\.?[0-9]+?$'
 
 
+class BlockShape(Enum):
+    """Enum class representing different block shapes
+    """
+
+    SQUARE = 'square'
+    """Square texture
+    """
+    SLAB = 'slab'
+    """Bottom half of the texture
+    """
+    STAIR = 'stair'
+    """All corners of the texture except the top right corner
+    """
+    CARPET = 'carpet'
+    """Only the bottom row of pixels on the texture
+    """
+    GLASS_PANE = 'glass_pane'
+    """Only the middle column of pixels on the texture
+    """
+
+
 class Edition(ABC):
+    """Base class for Minecraft editions.
+    """
 
     @abstractmethod
     def get_textures(self,
                      version_or_type: VersionType | str,
                      output_dir: str = DEFAULTS['OUTPUT_DIR'],
-                     scale_factor: int = DEFAULTS['SCALE_FACTOR'],
-                     do_merge: bool = DEFAULTS['DO_MERGE']):
+                     options: TextureOptions | None = None) -> str | None:
         """Extract, filter, and scale item and block textures.
 
         Args:
             version_or_type (str): a Minecraft version type, or a version string.
-            output_dir (str, optional): directory that the final textures will go.
-            scale_factor (int, optional): factor that will be used to scale the textures.
-            do_merge (bool, optional): whether to merge the block and item textures into a single directory.
+            output_dir (str, optional): directory that the final textures will go
+            options (TextureOptions | None, optional): options for the textures
 
         Returns:
             string | None: path of the final textures or None if invalid input
@@ -55,9 +78,6 @@ class Edition(ABC):
 
         Args:
             version_type (VersionType): type of version to get
-
-        Raises:
-            Exception: if the version number is invalid
 
         Returns:
             str: latest version as a string
@@ -127,25 +147,6 @@ class Edition(ABC):
             re.match(REGEX_JAVA_RC, version))
 
     @staticmethod
-    def merge_dirs(input_dir: str, output_dir: str):
-        """Merges block and item textures to a single directory.
-        Item textures are given priority when there are conflicts.
-
-        Args:
-            input_dir (str): directory in which there are subdirectories 'block' and 'item'
-            output_dir (str): directory in which the files will be merged into
-        """
-
-        block_folder = f'{input_dir}/blocks'
-        item_folder = f'{input_dir}/items'
-
-        tabbed_print(texts.TEXTURES_MERGING)
-        copytree(block_folder, output_dir, dirs_exist_ok=True)
-        rmtree(block_folder)
-        copytree(item_folder, output_dir, dirs_exist_ok=True)
-        rmtree(item_folder)
-
-    @staticmethod
     def filter_unwanted(input_dir: str,
                         output_dir: str,
                         edition: EditionType = EditionType.JAVA) -> str:
@@ -154,7 +155,7 @@ class Edition(ABC):
         Args:
             input_path (str): directory where the input files are
             output_path (str): directory where accepted files will end up
-            edition (EditionType, optional): type of edition, defaults to `EditionType.JAVA`
+            edition (EditionType, optional): type of edition
         """
 
         mk_dir(output_dir, del_prev=True)
@@ -174,17 +175,97 @@ class Edition(ABC):
         return output_dir
 
     @staticmethod
-    def scale_textures(path: str,
-                       scale_factor: int = 100,
-                       do_merge: bool = True,
-                       crop: bool = True) -> str:
+    def crop_texture(image_path: str,
+                     crop_shape: BlockShape,
+                     output_path: str | None = None):
+        """Crops a texture to a specific shape.
+
+        Args:
+            image_path (str): path of the texture to crop
+            crop_shape (BlockShape): shape to crop the texture to
+            output_path (str, optional): path to save the cropped texture to
+        """
+
+        if output_path is None:
+            output_path = image_path
+
+        transparent_color = (255, 255, 255, 0)
+
+        match crop_shape:
+            case BlockShape.SQUARE:
+                pil_image.open(image_path).crop(
+                    (0, 0, 16, 16)).save(output_path)
+
+            case BlockShape.SLAB:
+                img = pil_image.open(image_path).convert("RGBA")
+                img.paste(transparent_color, (0, 0, 16, 8))
+                img.save(output_path)
+
+            case BlockShape.STAIR:
+                img = pil_image.open(image_path).convert("RGBA")
+                img.paste(transparent_color, (8, 0, 16, 8))
+                img.save(output_path)
+
+            case BlockShape.CARPET:
+                img = pil_image.open(image_path).convert("RGBA")
+                img.paste(transparent_color, (0, 0, 16, 15))
+                img.save(output_path)
+
+            case BlockShape.GLASS_PANE:
+                img = pil_image.open(image_path).convert("RGBA")
+                img.paste(transparent_color, (0, 0, 7, 16))
+                img.paste(transparent_color, (9, 0, 16, 16))
+                img.save(output_path)
+
+            case _:
+                raise ValueError(f'Unknown block shape {crop_shape}')
+
+    @staticmethod
+    def replicate_textures(asset_dir: str, replication_rules: dict[str,
+                                                                   str]) -> int:
+        """Replicates textures in a directory
+
+        Args:
+            asset_dir (str): path to the directory containing the textures
+            replication_rules (dict): dictionary containing the replication rules
+
+        Returns:
+            int: number of textures replicated
+        """
+
+        tabbed_print(texts.TEXTURES_REPLICATING)
+
+        count = 0
+        originals = list(replication_rules.keys())
+        for subdir, _, files in os.walk(asset_dir):
+            for fil in files:
+                file_name = fil.replace('.png', '')
+                if file_name in originals:
+                    original_path = os.path.normpath(
+                        f"{os.path.abspath(subdir)}/{fil}")
+                    replicated_path = os.path.normpath(
+                        f"{os.path.abspath(subdir)}/{replication_rules[file_name]}.png"
+                    )
+                    copyfile(original_path, replicated_path)
+                    Edition.crop_texture(replicated_path, BlockShape.GLASS_PANE,
+                                         replicated_path)
+                    count += 1
+
+        return count
+
+    @staticmethod
+    def scale_textures(
+            path: str,
+            scale_factor: int = DEFAULTS['TEXTURE_OPTIONS']['SCALE_FACTOR'],
+            do_merge: bool = DEFAULTS['TEXTURE_OPTIONS']['DO_MERGE'],
+            do_crop: bool = DEFAULTS['TEXTURE_OPTIONS']['DO_CROP']) -> str:
         """Scales textures within a directory by a factor
 
         Args:
             path (str): path of the textures that will be scaled
             scale_factor (int, optional): factor that the textures will be scaled by
             do_merge (bool, optional): whether to merge block and item texture files into a single directory
-            crop (bool, optional): whether to crop non-square textures to be square
+            do_crop (bool, optional): whether to crop non-square textures to be square
 
         Returns:
             string: path of the scaled textures
@@ -193,30 +274,52 @@ class Edition(ABC):
         if do_merge:
             Edition.merge_dirs(path, path)
         tabbed_print(texts.TEXTURES_FILTERING)
+
+        if scale_factor == 1:
+            return path
+
         for subdir, _, files in os.walk(path):
             f.filter(f'{os.path.abspath(subdir)}', ['.png'])
 
-            if scale_factor != 1 and len(files) > 0:
-                if do_merge:
-                    tabbed_print(
-                        texts.TEXTURES_RESIZING_AMOUNT.format(
-                            texture_amount=len(files)))
-                else:
-                    tabbed_print(
-                        texts.TEXTURES_RESISING_AMOUNT_IN_DIR.format(
-                            texture_amount=len(files),
-                            dir_name=os.path.basename(subdir)))
+            if len(files) == 0:
+                continue
+
+            if do_merge:
+                tabbed_print(
+                    texts.TEXTURES_RESIZING_AMOUNT.format(
+                        texture_amount=len(files)))
+            else:
+                tabbed_print(
+                    texts.TEXTURES_RESISING_AMOUNT_IN_DIR.format(
+                        texture_amount=len(files),
+                        dir_name=os.path.basename(subdir)))
 
             for fil in files:
                 image_path = os.path.normpath(
                     f"{os.path.abspath(subdir)}/{fil}")
-                if crop:
-                    with pil_image.open(image_path) as img:
-                        if img.size[0] > 16 or img.size[1] > 16:
-                            img = img.crop((0, 0, 16, 16))
-                            img.save(image_path)
+                if do_crop:
+                    Edition.crop_texture(image_path, BlockShape.SQUARE,
+                                         image_path)
 
-                if scale_factor != 1:
-                    image.scale(image_path, scale_factor, scale_factor)
+                image.scale(image_path, scale_factor, scale_factor)
 
         return path
+
+    @staticmethod
+    def merge_dirs(input_dir: str, output_dir: str):
+        """Merges block and item textures to a single directory.
+        Item textures are given priority when there are conflicts.
+
+        Args:
+            input_dir (str): directory in which there are subdirectories 'block' and 'item'
+            output_dir (str): directory in which the files will be merged into
+        """
+
+        block_folder = f'{input_dir}/blocks'
+        item_folder = f'{input_dir}/items'
+
+        tabbed_print(texts.TEXTURES_MERGING)
+        copytree(block_folder, output_dir, dirs_exist_ok=True)
+        rmtree(block_folder)
+        copytree(item_folder, output_dir, dirs_exist_ok=True)
+        rmtree(item_folder)
