@@ -16,6 +16,7 @@ STYLED_TAB = style(f"{' '*4}* ", fg=Fg.CYAN)
 COMPLETED = style("Completed. You can find the textures on:\n{output_dir}\n",
                   fg=Fg.GREEN)
 CLEARING_TEMP = "Clearing temporary files..."
+CREATING_PARTIALS = "Creating partial textures..."
 EDITION_USING_X = "Using {edition} Edition."
 ERROR_COMMAND_FAILED = "The command failed with return code {error_code}: {error_msg}!"
 ERROR_EDITION_INVALID = "Invalid edition!"
@@ -25,8 +26,9 @@ FILES_DOWNLOADING = "Downloading assets..."
 FILES_EXTRACTING_N = "Extracting {file_amount} files..."
 TEXTURES_FILTERING = "Filtering textures..."
 TEXTURES_MERGING = "Merging block and item textures to a single directory..."
-TEXTURES_RESIZING_AMOUNT = "Resizing {texture_amount} textures..."
+TEXTURES_REPLICATING = "Replicating textures..."
 TEXTURES_RESISING_AMOUNT_IN_DIR = "Resizing {texture_amount} {dir_name} textures..."
+TEXTURES_RESIZING_AMOUNT = "Resizing {texture_amount} textures..."
 VERSION_LATEST_FINDING = "Finding latest version from {version_type} releases channel..."
 VERSION_LATEST_IS = "Latest version from {version_type} releases channel is {latest_version}."
 VERSION_USING_X = "Using {version} version."
